# frontend/openapi.json
{"openapi":"3.0.1","info":{"title":"Todo App API","description":"Spring Boot + Supabase Todo REST API","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Todo","description":"Todo CRUD API (JWT 인증 필요)"},{"name":"Auth","description":"회원가입 및 로그인 API"}],"paths":{"/api/v1/todos":{"get":{"tags":["Todo"],"summary":"Todo 전체 조회","description":"로그인한 사용자의 Todo 목록을 반환합니다.","operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTodoResponse"}}}}}},"post":{"tags":["Todo"],"summary":"Todo 생성","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoCreateRequest"}}},"required":true},"responses":{"201":{"description":"생성 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTodoResponse"}}}},"400":{"description":"유효하지 않은 요청","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTodoResponse"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Auth"],"summary":"회원가입","description":"이메일과 비밀번호로 새 계정을 생성합니다.","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"400":{"description":"유효하지 않은 요청 또는 이미 사용 중인 이메일","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}},"201":{"description":"회원가입 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"로그인","description":"이메일과 비밀번호로 로그인하여 JWT를 발급받습니다.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"로그인 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}},"400":{"description":"이메일 또는 비밀번호 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}},"/api/v1/todos/{id}":{"get":{"tags":["Todo"],"summary":"Todo 단건 조회","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"존재하지 않는 Todo","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTodoResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTodoResponse"}}}}}},"delete":{"tags":["Todo"],"summary":"Todo 삭제","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"삭제 성공"}}},"patch":{"tags":["Todo"],"summary":"Todo 수정 (부분)","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTodoResponse"}}}}}}}},"components":{"schemas":{"TodoCreateRequest":{"required":["title"],"type":"object","properties":{"title":{"type":"string","description":"할 일 제목","example":"장보기"}}},"ApiResponseTodoResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/TodoResponse"},"message":{"type":"string"}}},"TodoResponse":{"type":"object","properties":{"id":{"type":"string","description":"Todo ID (UUID)","example":"550e8400-e29b-41d4-a716-446655440000"},"title":{"type":"string","description":"할 일 제목","example":"장보기"},"completed":{"type":"boolean","description":"완료 여부","example":false},"createdAt":{"type":"string","description":"생성 시각 (ISO 8601)","example":"2024-01-15T10:30:00Z"}},"description":"Todo 응답"},"SignupRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"}}},"ApiResponseAuthResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/AuthResponse"},"message":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string","description":"JWT 액세스 토큰"},"tokenType":{"type":"string","description":"토큰 타입","example":"bearer"},"expiresIn":{"type":"integer","description":"만료 시간 (초)","format":"int32","example":3600},"userId":{"type":"string","description":"사용자 UUID"},"email":{"type":"string","description":"이메일","example":"user@example.com"},"needsEmailConfirmation":{"type":"boolean","description":"이메일 인증 필요 여부","example":false}},"description":"인증 응답"},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"TodoUpdateRequest":{"type":"object","properties":{"title":{"type":"string","description":"수정할 제목"},"completed":{"type":"boolean","description":"완료 여부"}}},"ApiResponseListTodoResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TodoResponse"}},"message":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}